separate get_data_from_csv test from search page test. add search page initial configuration app test
