+++ test/test_search_page.py
@@ -1,10 +1,11 @@
 """
-This file contains the unit tests for the getDataFromCSV function 
+This file contains the unit tests for the 2_Search_Page.py streamlit page 
 """
 
 import unittest
 
 import pandas as pd
+from streamlit.testing.v1 import AppTest
 
 from ui.helpers.get_data_from_csv import get_data_from_csv
 
@@ -13,20 +14,19 @@ class TestSearchPage(unittest.TestCase):
     This class contains the unit tests for the getDataFromCSV function
     """
 
-    def test_get_data_from_csv_good_file_path(self):
-        """function to test the getDataFromCSV function
-        with valid file path
+    def test_display_title(self):
         """
-
-        actual = get_data_from_csv('data/landmarks_washington_full.csv')
-        self.assertIsInstance(actual, pd.DataFrame, "Output is not a pandas DataFrame")
-
-    def test_get_data_from_csv_bad_file_path(self):
-        """function to test the getDataFromCSV function
-        with invalid file path
+        function to test the title of the search page
         """
-        actual = get_data_from_csv('../data/landmarks_washington.csv')
-        self.assertIsNone(actual, "Output is not None")
+        at = AppTest.from_file("ui/pages/2_Search_Page.py").run()
+        assert at.title[0].value == "Washington Landmarks Search"
+
+    def test_display_text_input(self):
+        """
+        function to test the appearance of text input of the search page
+        """
+        at = AppTest.from_file("ui/pages/2_Search_Page.py").run()
+        assert at.text_input[0].label == "Search for a landmark by name"
 
 if __name__ == '__main__':
     unittest.main()
